Watcher-Scripts/requirements.txt
paho-mqtt
requests
google-cloud-storage
dotenv